Refactor chip class and move independent chip into separate classes
