+++ packages/flutter/lib/material.dart
@@ -45,6 +45,10 @@ export 'src/material/checkbox.dart';
 export 'src/material/checkbox_list_tile.dart';
 export 'src/material/checkbox_theme.dart';
 export 'src/material/chip.dart';
+export 'src/material/chip_action.dart';
+export 'src/material/chip_choice.dart';
+export 'src/material/chip_filter.dart';
+export 'src/material/chip_input.dart';
 export 'src/material/chip_theme.dart';
 export 'src/material/circle_avatar.dart';
 export 'src/material/color_scheme.dart';
