+++ packages/flutter/lib/src/material/chip_choice.dart
@@ -1,0 +1,193 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/widgets.dart';
+
+import 'chip.dart';
+import 'chip_theme.dart';
+import 'debug.dart';
+import 'theme_data.dart';
+
+/// A material design choice chip.
+///
+/// [ChoiceChip]s represent a single choice from a set. Choice chips contain
+/// related descriptive text or categories.
+///
+/// Requires one of its ancestors to be a [Material] widget. The [selected] and
+/// [label] arguments must not be null.
+///
+/// {@tool snippet}
+///
+/// ```dart
+/// class MyThreeOptions extends StatefulWidget {
+///   const MyThreeOptions({Key? key}) : super(key: key);
+///
+///   @override
+///   State<MyThreeOptions> createState() => _MyThreeOptionsState();
+/// }
+///
+/// class _MyThreeOptionsState extends State<MyThreeOptions> {
+///   int? _value = 1;
+///
+///   @override
+///   Widget build(BuildContext context) {
+///     return Wrap(
+///       children: List<Widget>.generate(
+///         3,
+///         (int index) {
+///           return ChoiceChip(
+///             label: Text('Item $index'),
+///             selected: _value == index,
+///             onSelected: (bool selected) {
+///               setState(() {
+///                 _value = selected ? index : null;
+///               });
+///             },
+///           );
+///         },
+///       ).toList(),
+///     );
+///   }
+/// }
+/// ```
+/// {@end-tool}
+///
+/// See also:
+///
+///  * [Chip], a chip that displays information and can be deleted.
+///  * [InputChip], a chip that represents a complex piece of information, such
+///    as an entity (person, place, or thing) or conversational text, in a
+///    compact form.
+///  * [FilterChip], uses tags or descriptive words as a way to filter content.
+///  * [ActionChip], represents an action related to primary content.
+///  * [CircleAvatar], which shows images or initials of people.
+///  * [Wrap], A widget that displays its children in multiple horizontal or
+///    vertical runs.
+///  * <https://material.io/design/components/chips.html>
+class ChoiceChip extends StatelessWidget
+    implements
+        ChipAttributes,
+        SelectableChipAttributes,
+        DisabledChipAttributes {
+  /// Create a chip that acts like a radio button.
+  ///
+  /// The [label], [selected], [autofocus], and [clipBehavior] arguments must
+  /// not be null. The [pressElevation] and [elevation] must be null or
+  /// non-negative. Typically, [pressElevation] is greater than [elevation].
+  const ChoiceChip({
+    Key? key,
+    this.avatar,
+    required this.label,
+    this.labelStyle,
+    this.labelPadding,
+    this.onSelected,
+    this.pressElevation,
+    required this.selected,
+    this.selectedColor,
+    this.disabledColor,
+    this.tooltip,
+    this.side,
+    this.shape,
+    this.clipBehavior = Clip.none,
+    this.focusNode,
+    this.autofocus = false,
+    this.backgroundColor,
+    this.padding,
+    this.visualDensity,
+    this.materialTapTargetSize,
+    this.elevation,
+    this.shadowColor,
+    this.selectedShadowColor,
+    this.avatarBorder = const CircleBorder(),
+  }) : assert(selected != null),
+       assert(label != null),
+       assert(clipBehavior != null),
+       assert(autofocus != null),
+       assert(pressElevation == null || pressElevation >= 0.0),
+       assert(elevation == null || elevation >= 0.0),
+       super(key: key);
+
+  @override
+  final Widget? avatar;
+  @override
+  final Widget label;
+  @override
+  final TextStyle? labelStyle;
+  @override
+  final EdgeInsetsGeometry? labelPadding;
+  @override
+  final ValueChanged<bool>? onSelected;
+  @override
+  final double? pressElevation;
+  @override
+  final bool selected;
+  @override
+  final Color? disabledColor;
+  @override
+  final Color? selectedColor;
+  @override
+  final String? tooltip;
+  @override
+  final BorderSide? side;
+  @override
+  final OutlinedBorder? shape;
+  @override
+  final Clip clipBehavior;
+  @override
+  final FocusNode? focusNode;
+  @override
+  final bool autofocus;
+  @override
+  final Color? backgroundColor;
+  @override
+  final EdgeInsetsGeometry? padding;
+  @override
+  final VisualDensity? visualDensity;
+  @override
+  final MaterialTapTargetSize? materialTapTargetSize;
+  @override
+  final double? elevation;
+  @override
+  final Color? shadowColor;
+  @override
+  final Color? selectedShadowColor;
+  @override
+  final ShapeBorder avatarBorder;
+
+  @override
+  bool get isEnabled => onSelected != null;
+
+  @override
+  Widget build(BuildContext context) {
+    assert(debugCheckHasMaterial(context));
+    final ChipThemeData chipTheme = ChipTheme.of(context);
+    return RawChip(
+      avatar: avatar,
+      label: label,
+      labelStyle: labelStyle ?? (selected ? chipTheme.secondaryLabelStyle : null),
+      labelPadding: labelPadding,
+      onSelected: onSelected,
+      pressElevation: pressElevation,
+      selected: selected,
+      showCheckmark: false,
+      tooltip: tooltip,
+      side: side,
+      shape: shape,
+      clipBehavior: clipBehavior,
+      focusNode: focusNode,
+      autofocus: autofocus,
+      disabledColor: disabledColor,
+      selectedColor: selectedColor ?? chipTheme.secondarySelectedColor,
+      backgroundColor: backgroundColor,
+      padding: padding,
+      visualDensity: visualDensity,
+      isEnabled: isEnabled,
+      materialTapTargetSize: materialTapTargetSize,
+      elevation: elevation,
+      shadowColor: shadowColor,
+      selectedShadowColor: selectedShadowColor,
+      avatarBorder: avatarBorder,
+    );
+  }
+}
