+++ packages/flutter/test/material/chip_test.dart
@@ -133,62 +133,6 @@ Future<void> _testConstrainedLabel(
   expect(chipSize.height, chipParentHeight);
 }
 
-Widget _selectedInputChip({ Color? checkmarkColor }) {
-  return InputChip(
-    label: const Text('InputChip'),
-    selected: true,
-    showCheckmark: true,
-    checkmarkColor: checkmarkColor,
-  );
-}
-
-Widget _selectedFilterChip({ Color? checkmarkColor }) {
-  return FilterChip(
-    label: const Text('InputChip'),
-    selected: true,
-    showCheckmark: true,
-    checkmarkColor: checkmarkColor,
-    onSelected: (bool _) { },
-  );
-}
-
-Future<void> _pumpCheckmarkChip(
-  WidgetTester tester, {
-  required Widget chip,
-  Color? themeColor,
-  Brightness brightness = Brightness.light,
-}) async {
-  await tester.pumpWidget(
-    _wrapForChip(
-      brightness: brightness,
-      child: Builder(
-        builder: (BuildContext context) {
-          final ChipThemeData chipTheme = ChipTheme.of(context);
-          return ChipTheme(
-            data: themeColor == null ? chipTheme : chipTheme.copyWith(
-              checkmarkColor: themeColor,
-            ),
-            child: chip,
-          );
-        },
-      ),
-    ),
-  );
-}
-
-void _expectCheckmarkColor(Finder finder, Color color) {
-  expect(
-    finder,
-    paints
-      // The first path that is painted is the selection overlay. We do not care
-      // how it is painted but it has to be added it to this pattern so that the
-      // check mark can be checked next.
-      ..path()
-      // The second path that is painted is the check mark.
-      ..path(color: color),
-  );
-}
-
 void _doNothing() {}
 
 Widget _chipWithOptionalDeleteButton({
@@ -345,39 +289,6 @@ void main() {
     expect(labelStyle.overflow, textTheme.bodyText1?.overflow);
     expect(labelStyle.textBaseline, textTheme.bodyText1?.textBaseline);
     expect(labelStyle.wordSpacing, textTheme.bodyText1?.wordSpacing);
-  });
-
-  testWidgets('ChoiceChip defaults', (WidgetTester tester) async {
-    Widget buildFrame(Brightness brightness) {
-      return MaterialApp(
-        theme: ThemeData(brightness: brightness),
-        home: const Scaffold(
-          body: Center(
-            child: ChoiceChip(
-              label: Text('Chip A'),
-              selected: true,
-            ),
-          ),
-        ),
-      );
-    }
-
-    await tester.pumpWidget(buildFrame(Brightness.light));
-    expect(getMaterialBox(tester), paints..path(color: const Color(0x3d000000)));
-    expect(tester.getSize(find.byType(ChoiceChip)), const Size(108.0, 48.0));
-    expect(getMaterial(tester).color, null);
-    expect(getMaterial(tester).elevation, 0);
-    expect(getMaterial(tester).shape, const StadiumBorder());
-    expect(getLabelStyle(tester, 'Chip A').style.color?.value, 0xde000000);
-
-    await tester.pumpWidget(buildFrame(Brightness.dark));
-    await tester.pumpAndSettle(); // Theme transition animation
-    expect(getMaterialBox(tester), paints..path(color: const Color(0x3dffffff)));
-    expect(tester.getSize(find.byType(ChoiceChip)), const Size(108.0, 48.0));
-    expect(getMaterial(tester).color, null);
-    expect(getMaterial(tester).elevation, 0);
-    expect(getMaterial(tester).shape, const StadiumBorder());
-    expect(getLabelStyle(tester, 'Chip A').style.color?.value, 0xdeffffff);
   });
 
   testWidgets('Chip control test', (WidgetTester tester) async {
@@ -2477,20 +2388,6 @@ void main() {
     await tester.pumpWidget(
       const MaterialApp(
         home: Material(
-          child: ChoiceChip(
-            selected: false,
-            label: Text('choice chip'),
-          ),
-        ),
-      ),
-    );
-
-    await tester.tap(find.byType(ChoiceChip));
-    expect(tester.takeException(), null);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Material(
           child: RawChip(
             label: Text('raw chip'),
           ),
@@ -2499,47 +2396,6 @@ void main() {
     );
 
     await tester.tap(find.byType(RawChip));
-    expect(tester.takeException(), null);
-
-    await tester.pumpWidget(
-      MaterialApp(
-        home: Material(
-          child: ActionChip(
-            onPressed: () { },
-            label: const Text('action chip'),
-          ),
-        ),
-      ),
-    );
-
-    await tester.tap(find.byType(ActionChip));
-    expect(tester.takeException(), null);
-
-    await tester.pumpWidget(
-      MaterialApp(
-        home: Material(
-          child: FilterChip(
-            onSelected: (bool valueChanged) { },
-            label: const Text('filter chip'),
-          ),
-        ),
-      ),
-    );
-
-    await tester.tap(find.byType(FilterChip));
-    expect(tester.takeException(), null);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Material(
-          child: InputChip(
-            label: Text('input chip'),
-          ),
-        ),
-      ),
-    );
-
-    await tester.tap(find.byType(InputChip));
     expect(tester.takeException(), null);
   });
 
@@ -2645,42 +2501,6 @@ void main() {
     checkChipMaterialClipBehavior(tester, Clip.none);
 
     await tester.pumpWidget(_wrapForChip(child: const Chip(label: label, clipBehavior: Clip.antiAlias)));
-    checkChipMaterialClipBehavior(tester, Clip.antiAlias);
-  });
-
-  testWidgets('ChoiceChip clipBehavior properly passes through to the Material', (WidgetTester tester) async {
-    const Text label = Text('label');
-    await tester.pumpWidget(_wrapForChip(child: const ChoiceChip(label: label, selected: false)));
-    checkChipMaterialClipBehavior(tester, Clip.none);
-
-    await tester.pumpWidget(_wrapForChip(child: const ChoiceChip(label: label, selected: false, clipBehavior: Clip.antiAlias)));
-    checkChipMaterialClipBehavior(tester, Clip.antiAlias);
-  });
-
-  testWidgets('FilterChip clipBehavior properly passes through to the Material', (WidgetTester tester) async {
-    const Text label = Text('label');
-    await tester.pumpWidget(_wrapForChip(child: FilterChip(label: label, onSelected: (bool b) { })));
-    checkChipMaterialClipBehavior(tester, Clip.none);
-
-    await tester.pumpWidget(_wrapForChip(child: FilterChip(label: label, onSelected: (bool b) { }, clipBehavior: Clip.antiAlias)));
-    checkChipMaterialClipBehavior(tester, Clip.antiAlias);
-  });
-
-  testWidgets('ActionChip clipBehavior properly passes through to the Material', (WidgetTester tester) async {
-    const Text label = Text('label');
-    await tester.pumpWidget(_wrapForChip(child: ActionChip(label: label, onPressed: () { })));
-    checkChipMaterialClipBehavior(tester, Clip.none);
-
-    await tester.pumpWidget(_wrapForChip(child: ActionChip(label: label, clipBehavior: Clip.antiAlias, onPressed: () { })));
-    checkChipMaterialClipBehavior(tester, Clip.antiAlias);
-  });
-
-  testWidgets('InputChip clipBehavior properly passes through to the Material', (WidgetTester tester) async {
-    const Text label = Text('label');
-    await tester.pumpWidget(_wrapForChip(child: const InputChip(label: label)));
-    checkChipMaterialClipBehavior(tester, Clip.none);
-
-    await tester.pumpWidget(_wrapForChip(child: const InputChip(label: label, clipBehavior: Clip.antiAlias)));
     checkChipMaterialClipBehavior(tester, Clip.antiAlias);
   });
 
@@ -3198,71 +3018,6 @@ void main() {
     expect(find.byType(RawChip), paints..drrect(color: selectedBorderSide.color));
   });
 
-  testWidgets('loses focus when disabled', (WidgetTester tester) async {
-    final FocusNode focusNode = FocusNode(debugLabel: 'InputChip');
-    await tester.pumpWidget(
-      _wrapForChip(
-        child: InputChip(
-          focusNode: focusNode,
-          autofocus: true,
-          shape: const RoundedRectangleBorder(),
-          avatar: const CircleAvatar(child: Text('A')),
-          label: const Text('Chip A'),
-          onPressed: () { },
-        ),
-      ),
-    );
-    await tester.pump();
-    expect(focusNode.hasPrimaryFocus, isTrue);
-
-    await tester.pumpWidget(
-      _wrapForChip(
-        child: InputChip(
-          focusNode: focusNode,
-          autofocus: true,
-          shape: const RoundedRectangleBorder(),
-          avatar: const CircleAvatar(child: Text('A')),
-          label: const Text('Chip A'),
-        ),
-      ),
-    );
-    await tester.pump();
-    expect(focusNode.hasPrimaryFocus, isFalse);
-  });
-
-  testWidgets('cannot be traversed to when disabled', (WidgetTester tester) async {
-    final FocusNode focusNode1 = FocusNode(debugLabel: 'InputChip 1');
-    final FocusNode focusNode2 = FocusNode(debugLabel: 'InputChip 2');
-    await tester.pumpWidget(
-      _wrapForChip(
-        child: Column(
-          children: <Widget>[
-            InputChip(
-              focusNode: focusNode1,
-              autofocus: true,
-              label: const Text('Chip A'),
-              onPressed: () { },
-            ),
-            InputChip(
-              focusNode: focusNode2,
-              autofocus: true,
-              label: const Text('Chip B'),
-            ),
-          ],
-        ),
-      ),
-    );
-    await tester.pump();
-    expect(focusNode1.hasPrimaryFocus, isTrue);
-    expect(focusNode2.hasPrimaryFocus, isFalse);
-
-    expect(focusNode1.nextFocus(), isTrue);
-
-    await tester.pump();
-    expect(focusNode1.hasPrimaryFocus, isTrue);
-    expect(focusNode2.hasPrimaryFocus, isFalse);
-  });
-
   testWidgets('Chip responds to density changes.', (WidgetTester tester) async {
     const Key key = Key('test');
     const Key textKey = Key('test text');
@@ -3366,132 +3121,6 @@ void main() {
     await tester.pumpAndSettle();
     box = tester.getRect(find.byKey(key));
     expect(box.size, equals(const Size(128, 24.0 + 16.0)));
-  });
-
-  testWidgets('Input chip check mark color is determined by platform brightness when light', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedInputChip(),
-    );
-
-    _expectCheckmarkColor(
-      find.byType(InputChip),
-      Colors.black.withAlpha(0xde),
-    );
-  });
-
-  testWidgets('Filter chip check mark color is determined by platform brightness when light', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedFilterChip(),
-    );
-
-    _expectCheckmarkColor(
-      find.byType(FilterChip),
-      Colors.black.withAlpha(0xde),
-    );
-  });
-
-  testWidgets('Input chip check mark color is determined by platform brightness when dark', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedInputChip(),
-      brightness: Brightness.dark,
-    );
-
-    _expectCheckmarkColor(
-      find.byType(InputChip),
-      Colors.white.withAlpha(0xde),
-    );
-  });
-
-  testWidgets('Filter chip check mark color is determined by platform brightness when dark', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedFilterChip(),
-      brightness: Brightness.dark,
-    );
-
-    _expectCheckmarkColor(
-      find.byType(FilterChip),
-      Colors.white.withAlpha(0xde),
-    );
-  });
-
-  testWidgets('Input chip check mark color can be set by the chip theme', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedInputChip(),
-      themeColor: const Color(0xff00ff00),
-    );
-
-    _expectCheckmarkColor(
-      find.byType(InputChip),
-      const Color(0xff00ff00),
-    );
-  });
-
-  testWidgets('Filter chip check mark color can be set by the chip theme', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedFilterChip(),
-      themeColor: const Color(0xff00ff00),
-    );
-
-    _expectCheckmarkColor(
-      find.byType(FilterChip),
-      const Color(0xff00ff00),
-    );
-  });
-
-  testWidgets('Input chip check mark color can be set by the chip constructor', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedInputChip(checkmarkColor: const Color(0xff00ff00)),
-    );
-
-    _expectCheckmarkColor(
-      find.byType(InputChip),
-      const Color(0xff00ff00),
-    );
-  });
-
-  testWidgets('Filter chip check mark color can be set by the chip constructor', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedFilterChip(checkmarkColor: const Color(0xff00ff00)),
-    );
-
-    _expectCheckmarkColor(
-      find.byType(FilterChip),
-      const Color(0xff00ff00),
-    );
-  });
-
-  testWidgets('Input chip check mark color is set by chip constructor even when a theme color is specified', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedInputChip(checkmarkColor: const Color(0xffff0000)),
-      themeColor: const Color(0xff00ff00),
-    );
-
-    _expectCheckmarkColor(
-      find.byType(InputChip),
-      const Color(0xffff0000),
-    );
-  });
-
-  testWidgets('Filter chip check mark color is set by chip constructor even when a theme color is specified', (WidgetTester tester) async {
-    await _pumpCheckmarkChip(
-      tester,
-      chip: _selectedFilterChip(checkmarkColor: const Color(0xffff0000)),
-      themeColor: const Color(0xff00ff00),
-    );
-
-    _expectCheckmarkColor(
-      find.byType(FilterChip),
-      const Color(0xffff0000),
-    );
   });
 
   testWidgets('Chip delete button tooltip can be disabled using useDeleteButtonTooltip', (WidgetTester tester) async {
